Scrapers now use SET_SHOES by default.
When new shoes are found the scraper sends them to the monitor via SET_SHOES.
The monitor then checks for duplicate links, adding only actually new shoes.

This is done because many times, if the scraper is restarted,
it takes a long time for it to find all the other links,
possibly even missing them.
Also, this way the monitor does not lose any of the old links
while still getting the new ones.