Added on_config_change endpoint.
This is done so that self.shoes can be modified before the actual loop,
checking that all shoes should be monitored according to the new config
and removing the ones that don't belong.
The function itself is not implemented since the type of parameters
to check will depend on the monitor/scraper itself.